some commenting, fixed a bug in shoppingcart, minor changes to other classes

+++ src/Item.java
@@ -5,7 +5,7 @@ public abstract class Item{
 		
 		protected int sNo;
 		protected String title;
-		protected int price;
+		protected double price;
 		protected int quantity;
 		protected String itemtype;
 //		public abstract String[] getInfo(String filename, int lineNumber);
